+++ hivemq-extension-sdk/src/main/java/com/hivemq/extension/sdk/api/packets/pubrel/PubrelPacket.java
@@ -15,8 +15,8 @@ import java.util.Optional;
  *
  * @author Yannick Weber
  */
-@DoNotImplement
 @Immutable
+@DoNotImplement
 public interface PubrelPacket {
 
     /**
