+++ src/main/java/com/hivemq/extensions/packets/pubrel/PubrelPacketImpl.java
@@ -1,8 +1,8 @@
 package com.hivemq.extensions.packets.pubrel;
 
-import com.hivemq.annotations.Nullable;
 import com.hivemq.extension.sdk.api.annotations.Immutable;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
+import com.hivemq.extension.sdk.api.annotations.Nullable;
 import com.hivemq.extension.sdk.api.packets.general.UserProperties;
 import com.hivemq.extension.sdk.api.packets.publish.PubrelReasonCode;
 import com.hivemq.extension.sdk.api.packets.pubrel.PubrelPacket;
